# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger KW-API",
        "description": "This is KW-API messaging middleware for dispatch data to services / applications",
        "contact": {
            "email": "izhur2001@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/v1",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/events": {
            "get": {
                "tags": [
                    "events"
                ],
                "summary": "list event data",
                "description": "",
                "operationId": "list event",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page data",
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiKey": [

                        ]
                    }
                ]
            }
        },
        "/events/{eventId}": {
            "get": {
                "tags": [
                    "events"
                ],
                "summary": "show detail of event data",
                "description": "",
                "operationId": "show event",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "id of event",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiKey": [

                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "events"
                ],
                "summary": "update event data",
                "description": "",
                "operationId": "update event",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "id of event",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "object",
                        "in": "formData",
                        "description": "object of event",
                        "required": true,
                        "type": "string",
                        "default": "lead"
                    },
                    {
                        "name": "action",
                        "in": "formData",
                        "description": "action of event",
                        "required": true,
                        "type": "string",
                        "default": "create"
                    },
                    {
                        "name": "version",
                        "in": "formData",
                        "description": "version of event",
                        "required": true,
                        "type": "string",
                        "default": "1"
                    },
                    {
                        "name": "jsonSchema",
                        "in": "formData",
                        "description": "json schema of data event that that needs to dispatch",
                        "required": true,
                        "type": "string",
                        "default": ""
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiKey": [

                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "events"
                ],
                "summary": "delete event data",
                "description": "",
                "operationId": "delete event",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "id of event",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiKey": [

                        ]
                    }
                ]
            }
        },
        "/events/unsubscribe": {
            "post": {
                "tags": [
                    "events"
                ],
                "summary": "unsubscribe event data",
                "description": "",
                "operationId": "unsubscribe event",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "object",
                        "in": "formData",
                        "description": "object of event",
                        "required": true,
                        "type": "string",
                        "default": "lead"
                    },
                    {
                        "name": "action",
                        "in": "formData",
                        "description": "action of event",
                        "required": true,
                        "type": "string",
                        "default": "create"
                    },
                    {
                        "name": "version",
                        "in": "formData",
                        "description": "version of event",
                        "required": true,
                        "type": "string",
                        "default": "1"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiKey": [

                        ]
                    }
                ]
            }
        },
        "/events/register": {
            "post": {
                "tags": [
                    "events"
                ],
                "summary": "Register an event from services / application publisher",
                "description": "",
                "operationId": "registerEvent",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "object",
                        "in": "formData",
                        "description": "object of event",
                        "required": true,
                        "type": "string",
                        "default": "lead"
                    },
                    {
                        "name": "action",
                        "in": "formData",
                        "description": "action of event",
                        "required": true,
                        "type": "string",
                        "default": "create"
                    },
                    {
                        "name": "version",
                        "in": "formData",
                        "description": "version of event",
                        "required": true,
                        "type": "string",
                        "default": "1"
                    },
                    {
                        "name": "jsonSchema",
                        "in": "formData",
                        "description": "json schema of data event that that needs to dispatch",
                        "required": true,
                        "type": "string",
                        "default": ""
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiKey": [

                        ]
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "apiKey",
            "in": "header"
        }
    },
...
}